[simulation] suppressed warning

+++ src/libraries/simulation_kernel/Entities/Objects/BuildableCapacity.cpp
@@ -179,7 +179,7 @@ unsigned int BuildableCapacity::GetDotationNumber( const TER_Localisation& locat
 // -----------------------------------------------------------------------------
 void BuildableCapacity::SetDensity( double density )
 {
-    nFullNbrDotation_ = ( density * 10000 ); // $$$$ LDC Density of 0.1 = 1 mine per 100 square meter (from TTA 702 \\install\Masa\simulation\projects\scipio\doc\V1.STAB\DOC MODELISATION\DOCTRINE et memento\GEN\Doctrine page 131)
+    nFullNbrDotation_ = static_cast< unsigned int >( density * 10000 ); // $$$$ LDC Density of 0.1 = 1 mine per 100 square meter (from TTA 702 \\install\Masa\simulation\projects\scipio\doc\V1.STAB\DOC MODELISATION\DOCTRINE et memento\GEN\Doctrine page 131)
 }
 
 // -----------------------------------------------------------------------------
